Replace poor-choice HTML tags with better semantic tags
In about me article, the headings are not heading anything.
To me, they semantically look more like a quick list of
information for that thing. Key word here is list. I instead
made an unordered list here.

Before that, the <b/> and <i/> tags didn't seem to have semantic
meaning here. The h2 element was already indicating why that
text was special and stood apart - so b tag is not necessary.
Furthermore, the i tag seems to be there for its stylistic value
which is a no-go. The i tag should be used for idiomatic text
which doesnt' seem to be the case here. I think we'll be able
to find it later for styling.

The div with the class of header was not good either. If it's a
heading, there are heading tags made for that. I gave this an
h2 tag instead.

The p tag is used for "Facts About Me" and seems out of place.
It is not a complete body of text for one idea or a grouping
of related content, as mdn would have it. I removed this wrapping
p-tag and left the bare, exposed h2 tag.