Clean up ids and classes
No classes or ids should imply a style - they are purely semantic.
So, in  "About Me" the article has class="green-background" and
this is bad. This is the only place this is used. If the author
of this site wants a green background in only that article, they
can find id="primary" and apply css there. I removed this class.

The id="quick-description" didn't seem specific enough. A lot of
elements could have a "quick-description" that has different
needs. I made this a more specific id by tying it semantically to
the about me section.

There were multiple uses of id="home" when ids should occcur no
more than once on any given page. It seemed less important in
the <html> tag, so I removed it from that tag.

The article id="my dog" is not good. There should not be a space
in an id. I replaced the space with a hyphen.

List items all with the same parent shouldn't have the same class.
I fixed this in the "Quick Facts" part of the page.