Fix up quick facts section with nested list and better whitespace
It seems like the author REALLY wanted "Quick Facts" to be a
heading but instead of using a heading tag used p and b tags.
So, I replaced those tags with an h3 tag and made the rest of
quick facts a nested list. I get the idea that we may have
new categories of facts later on here (e.g. family facts,
career facts, etc.). So, we can add more li later.

I also deleted some comments here and removed the spaces between
text and tags in the li elements in the quick facts part.