Fix bad URIs
In the navigation, every link should point to a folder, not an
indvidual html file. For the home link, assuming this index.html
file is the home page, we should reference the current directory.
When the server serves that directory, it serves the index.html.
All files should be set up this way.

For the resume page, nothing should be a higher directory than the
home page index.html. So this shouldn't need "..". Also, we can
afford a more meaningful URL path than "my-res". It costs nothing
to say "resume" instead. so, "./resume" is the new path here.

What is "randomstuff"? Exactly. I made this a url called
"interests".

In the img, NO ONE should EVER use a file from the local machine.
Absolutely never. To fix this, I just used example.com like we
were shown in lecture. Alternatively, we could put the img in this
repository and use a relative URI for it.