Swap ul out for section instead with quick facts
The list inside of "Facts About Me" felt like a list of sections.
So, I think it would just be better to use sections than an
unordered list of "sections". I noticed this after trying to add
space under ul tags and it doubled up with the nested list and
I realized there was a semantic difference that would also be
useful.